+++ src/main/java/com/ohgiraffers/hitechautoworks/res/service/ResService.java
@@ -1,0 +1,26 @@
+package com.ohgiraffers.hitechautoworks.res.service;
+
+import com.ohgiraffers.hitechautoworks.res.dao.ResMapper;
+import com.ohgiraffers.hitechautoworks.res.dto.ResDTO;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+
+@Service
+public class ResService {
+
+    @Autowired
+    private ResMapper resMapper;
+
+    public ResDTO findUserRes(int resCode) {
+        return resMapper.findUserRes(resCode);
+    }
+
+    public List<ResDTO> findAllres() {
+        return resMapper.findAllres();
+    }
+
+    public List<ResDTO> findCodeRes(int resCode) { return resMapper.findCodeRes(resCode);
+    }
+}
