+++ src/main/resources/mapper/accountmapper.xml
@@ -1,0 +1,34 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<!DOCTYPE mapper
+        PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+        "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
+<mapper namespace="com.ohgiraffers.hitechautoworks.account.dao.AccountMapper">
+
+
+    <resultMap id="user" type="com.ohgiraffers.hitechautoworks.auth.dto.UserDTO">
+        <id property="userCode" column="user_code"/>
+        <result property="userId" column="user_id"/>
+        <result property="password" column="password"/>
+        <result property="userName" column="user_name"/>
+        <result property="userRole" column="user_role"/>
+    </resultMap>
+
+    <select id="findAllUser" resultMap="user">
+        SELECT
+        *
+        FROM TBL_USER
+    </select>
+
+    <select id="findUserCode" resultMap="user">
+        SELECT * from tbl_user
+        where user_code = #{userCode}
+    </select>
+
+    <select id="findUserName" resultMap="user">
+        SELECT
+        *
+        FROM TBL_USER
+        WHERE user_name = #{userName}
+    </select>
+
+</mapper>
