+++ src/main/resources/mapper/repairmapper.xml
@@ -1,0 +1,23 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<!DOCTYPE mapper
+        PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+        "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
+<mapper namespace="com.ohgiraffers.hitechautoworks.repair.dao.RepairMapper">
+
+    <resultMap id="repair" type="com.ohgiraffers.hitechautoworks.repair.dto.RepairDTO">
+        <result property="content" column="repair_content"/>
+        <result property="date" column="repair_date"/>
+        <result property="status" column="repair_status"/>
+        <association property="resDTO" resultMap="res" />
+    </resultMap>
+
+    <select id="findAllRepair" resultMap="repair">
+        SELECT
+        a.res_code,
+        a.repair_date,
+        a.repair_content
+        FROM
+        tbl_repair a
+
+    </select>
+</mapper>
