+++ src/main/resources/mapper/resmapper.xml
@@ -1,0 +1,54 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<!DOCTYPE mapper
+        PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+        "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
+<mapper namespace="com.ohgiraffers.hitechautoworks.res.dao.ResMapper">
+
+    <resultMap id="res" type="com.ohgiraffers.hitechautoworks.res.dto.ResDTO">
+        <id property="code" column="res_code"/>
+        <result property="title" column="res_title"/>
+        <result property="option" column="res_option"/>
+        <result property="date" column="res_date"/>
+        <result property="extra" column="res_extra"/>
+        <association property="userDTO" javaType="com.ohgiraffers.hitechautoworks.auth.dto.UserDTO">
+            <id property="userCode" column="user_code"/>
+            <result property="userId" column="user_id"/>
+            <result property="password" column="password"/>
+            <result property="userName" column="user_name"/>
+            <result property="userRole" column="user_role"/>
+        </association>
+    </resultMap>
+
+    <select id="findAllres" resultMap="res">
+        select
+        a.res_code,
+        a.res_title,
+        a.res_option,
+        a.res_date,
+        a.res_extra,
+        b.user_name
+        from
+        tbl_res a
+        join tbl_user b on (a.user_code = b.user_code)
+    </select>
+
+    <select id="findUserRes" resultMap="res">
+        SELECT *
+        FROM tbl_res
+        WHERE res_code = #{resCode}
+    </select>
+
+    <select id="findCodeRes" resultMap="res">
+        SELECT
+        a.res_code,
+        a.res_title,
+        a.res_option,
+        a.res_date,
+        a.res_extra,
+        b.user_name
+        FROM tbl_res a
+        JOIN tbl_user b ON (a.user_code = b.user_code)
+        WHERE a.res_code = #{code}
+    </select>
+
+</mapper>
