+++ src/main/resources/mapper/usermapper.xml
@@ -36,6 +36,8 @@
         SET user_role = 'CUSTOMER'
         WHERE user_id = #{userId} AND user_check = #{checknumber}
     </update>
+
+
     <resultMap id="user" type="com.ohgiraffers.hitechautoworks.auth.dto.UserDTO">
         <id property="userCode" column="user_code"/>
         <result property="userId" column="user_id"/>
@@ -43,17 +45,6 @@
         <result property="userName" column="user_name"/>
         <result property="userRole" column="user_role"/>
     </resultMap>
-    <select id="findAllUser" resultMap="user">
-        SELECT
-        *
-        FROM TBL_USER
-    </select>
-
-
-    <select id="findUserCode" resultMap="user">
-        SELECT * from tbl_user
-        where user_code = #{userCode}
-    </select>
 
     <resultMap id="finduser" type="com.ohgiraffers.hitechautoworks.auth.dto.UserDTO">
         <id property="userNo" column="user_no"/>
@@ -62,152 +53,8 @@
         <result property="password" column="password"/>
         <result property="userRole" column="user_role"/>
     </resultMap>
-    <select id="findUserName" resultMap="user">
-        SELECT
-        *
-        FROM TBL_USER
-        WHERE user_name = #{userName}
-    </select>
-
-        <resultMap id="partReusltMap" type="com.ohgiraffers.hitechautoworks.auth.dto.PartDTO">
-        <id property="partCode" column="part_code"/>
-        <result property="partName" column="part_name"/>
-        <result property="partstock" column="part_stock"/>
-        <result property="partPrice" column="part_price"/>
-    </resultMap>
-
-    <select id="selectPartByCode" resultMap="partReusltMap">
-        SELECT *
-        FROM tbl_part
-        WHERE part_code = #{partCode}
-    </select>
-
-    <select id="selectAllPart" resultMap="partReusltMap">
-        SELECT *
-        FROM tbl_part
-    </select>
-
-    <select id="selectpart" resultMap="partReusltMap">
-        SELECT *
-        FROM tbl_part
-        where part_code = #{partCode}
-    </select>
-
-<resultMap id="partResult" type="com.ohgiraffers.hitechautoworks.auth.dto.PartDTO">
-        <id property="partCode" column="PART_CODE"/>
-        <result property="partName" column="PART_NAME"/>
-        <result property="partstock" column="PART_STOCK"/>
-        <result property="partPrice" column="PART_PRICE"/>
-
-    </resultMap>
-    <select id="partSearchBtPartName" resultMap="partResult">
-        SELECT
-              *
-        FROM
-            tbl_part
-        WHERE
-             part_name = #{ partName }
-    </select>
 
 
-    <update id="modifyPart">
-        UPDATE
-            tbl_part
-        SET
-            PART_NAME = #{  partName },
-            PART_STOCK= #{ partstock },
-            PART_PRICE = #{ partPrice }
-        WHERE
-            PART_CODE = #{ partCode }
-    </update>
-
-    <insert id="addPart">
-    INSERT INTO tbl_part(part_name, part_stock, part_price)
-    VALUES(
-        #{ partName },
-        #{ partstock },
-        #{ partPrice }
-        )
-
-    </insert>
-
-
-        <resultMap id="res" type="com.ohgiraffers.hitechautoworks.auth.dto.ResDTO">
-        <id property="code" column="res_code"/>
-        <result property="title" column="res_title"/>
-        <result property="option" column="res_option"/>
-        <result property="date" column="res_date"/>
-        <result property="extra" column="res_extra"/>
-        <association property="userDTO" javaType="com.ohgiraffers.hitechautoworks.auth.dto.UserDTO">
-            <id property="userCode" column="user_code"/>
-            <result property="userId" column="user_id"/>
-            <result property="password" column="password"/>
-            <result property="userName" column="user_name"/>
-            <result property="userRole" column="user_role"/>
-        </association>
-    </resultMap>
-
-    <select id="findAllres" resultMap="res">
-        select
-            a.res_code,
-            a.res_title,
-            a.res_option,
-            a.res_date,
-            a.res_extra,
-            b.user_name
-        from
-            tbl_res a
-        join tbl_user b on (a.user_code = b.user_code)
-    </select>
-    <select id="findUserRes" resultMap="res">
-        SELECT *
-        FROM tbl_res
-        WHERE res_code = #{resCode}
-    </select>
-
-    <select id="findCodeRes" resultMap="res">
-        SELECT
-            a.res_code,
-            a.res_title,
-            a.res_option,
-            a.res_date,
-            a.res_extra,
-            b.user_name
-        FROM tbl_res a
-        JOIN tbl_user b ON (a.user_code = b.user_code)
-        WHERE a.res_code = #{code}
-    </select>
-
-
-    <delete id="deletePart">
-        DELETE FROM
-                    tbl_part
-        WHERE
-             part_code = #{ partCode }
-    </delete>
-
-    <resultMap id="repair" type="com.ohgiraffers.hitechautoworks.auth.dto.RepairDTO">
-        <result property="content" column="repair_content"/>
-        <result property="date" column="repair_date"/>
-        <result property="status" column="repair_status"/>
-        <association property="resDTO" resultMap="res" />
-    </resultMap>
-
-    <select id="findAllRepair" resultMap="repair">
-        SELECT
-        a.res_code,
-        a.repair_date,
-        a.repair_content,
-        c.user_name
-
-        FROM
-            tbl_repair a
-        JOIN
-            tbl_res b ON a.res_code = b.res_code
-        JOIN
-            tbl_repair_workers c ON b.user_code = c.user_code
-
-    </select>
 
 
 </mapper>
